Rename json related functions and put them in common.text.converters
Move formatters.py to common.text.formatters.py and update references in modules.
