+++ lib/ansible/module_utils/common/text/formatters.py
