+++ lib/ansible/module_utils/facts/hardware/sunos.py
@@ -20,7 +20,7 @@ import re
 
 from ansible.module_utils.six.moves import reduce
 
-from ansible.module_utils.text.formatters import bytes_to_human
+from ansible.module_utils.common.text.formatters import bytes_to_human
 
 from ansible.module_utils.facts.utils import get_file_content, get_mount_size
