+++ lib/ansible/modules/cloud/docker/docker_container.py
@@ -923,7 +923,7 @@ import re
 import shlex
 from distutils.version import LooseVersion
 
-from ansible.module_utils.text.formatters import human_to_bytes
+from ansible.module_utils.common.text.formatters import human_to_bytes
 from ansible.module_utils.docker.common import (
     AnsibleDockerClient,
     DifferenceTracker,
