+++ test/units/executor/module_common/test_recursive_finder.py
@@ -47,21 +47,20 @@ MODULE_UTILS_BASIC_IMPORTS = frozenset((('_text',),
                                         ('common', '_json_compat'),
                                         ('common', 'collections'),
                                         ('common', 'file'),
-                                        ('common', 'collections'),
                                         ('common', 'parameters'),
                                         ('common', 'process'),
                                         ('common', 'sys_info'),
+                                        ('common', 'text', '__init__'),
+                                        ('common', 'text', 'converters'),
+                                        ('common', 'text', 'formatters'),
                                         ('common', 'validation'),
                                         ('common', '_utils'),
-                                        ('json_utils',),
                                         ('distro', '__init__'),
                                         ('distro', '_distro'),
                                         ('parsing', '__init__'),
                                         ('parsing', 'convert_bool'),
                                         ('pycompat24',),
                                         ('six', '__init__'),
-                                        ('text', '__init__'),
-                                        ('text', 'formatters'),
                                         ))
 
 MODULE_UTILS_BASIC_FILES = frozenset(('ansible/module_utils/_text.py',
@@ -77,17 +76,17 @@ MODULE_UTILS_BASIC_FILES = frozenset(('ansible/module_utils/_text.py',
                                       'ansible/module_utils/common/file.py',
                                       'ansible/module_utils/common/process.py',
                                       'ansible/module_utils/common/sys_info.py',
+                                      'ansible/module_utils/common/text/__init__.py',
+                                      'ansible/module_utils/common/text/converters.py',
+                                      'ansible/module_utils/common/text/formatters.py',
                                       'ansible/module_utils/common/validation.py',
                                       'ansible/module_utils/common/_utils.py',
                                       'ansible/module_utils/distro/__init__.py',
                                       'ansible/module_utils/distro/_distro.py',
-                                      'ansible/module_utils/json_utils.py',
                                       'ansible/module_utils/parsing/__init__.py',
                                       'ansible/module_utils/parsing/convert_bool.py',
                                       'ansible/module_utils/pycompat24.py',
                                       'ansible/module_utils/six/__init__.py',
-                                      'ansible/module_utils/text/__init__.py',
-                                      'ansible/module_utils/text/formatters.py',
                                       ))
 
 ONLY_BASIC_IMPORT = frozenset((('basic',),))
